v0.2.0 (#5)
* fix: コメント修正

* fix: レスポンスの変数名を変更

* refactor: 環境変数取得処理を別ファイルに分離

* refactor: 共通リクエスト処理を別ファイルに分離

* fix: ユーザ情報のスコープを修正

* refactor: ユーザ情報取得処理を別ファイルに分離

* docs: パッケージのバージョンを更新
